+++ connectors/kafka/kafka-connector/src/main/java/org/platformlambda/kafka/ServiceQuery.java
@@ -25,16 +25,14 @@ import org.platformlambda.core.system.ServerPersonality;
 import org.platformlambda.core.system.ServiceDiscovery;
 import org.platformlambda.core.util.Utility;
 
-import java.util.Date;
-import java.util.HashMap;
-import java.util.List;
-import java.util.Map;
+import java.util.*;
 
 public class ServiceQuery implements LambdaFunction {
 
     private static final String TYPE = ServiceDiscovery.TYPE;
     private static final String ROUTE = ServiceDiscovery.ROUTE;
     private static final String FIND = ServiceDiscovery.FIND;
+    private static final String SEARCH = ServiceDiscovery.SEARCH;
     private static final String DOWNLOAD = "download";
     private static final String INFO = "info";
 
@@ -77,6 +75,9 @@ public class ServiceQuery implements LambdaFunction {
             } else {
                 return exists(route);
             }
+
+        } else if (SEARCH.equals(headers.get(TYPE)) && headers.containsKey(ROUTE)) {
+            return ServiceRegistry.getInstances(headers.get(ROUTE));
 
         } else {
             throw new IllegalArgumentException("Usage: type=download, info or (type=find, route=route_name)");