+++ connectors/kafka/kafka-connector/src/main/java/org/platformlambda/kafka/ServiceRegistry.java
@@ -80,6 +80,14 @@ public class ServiceRegistry implements LambdaFunction {
         return new HashMap<>(routes);
     }
 
+    public static List<String> getInstances(String route) {
+        if (routes.containsKey(route)) {
+            return new ArrayList<>(routes.get(route).keySet());
+        } else {
+            return Collections.emptyList();
+        }
+    }
+
     public static Map<String, String> getAllOrigins() {
         return new HashMap<>(origins);
     }